+++ lib/src/models/crypto_book.dart
@@ -1,9 +1,14 @@
 class CryptoBook {
+  /// List of recent Bid
   final List<Price> bids;
+
+  /// List of recent Ask
   final List<Price> asks;
 
+  /// Default Constructor
   CryptoBook(this.bids, this.asks);
 
+  /// Convert JSON to Object
   static CryptoBook? parse(Map<String, dynamic>? data) {
     if (data == null) {
       return null;
@@ -17,8 +22,13 @@ class CryptoBook {
 }
 
 class Price {
+  /// The price of the bid or ask
   final num price;
+
+  /// The total quantity remaining at the price
   final num size;
+
+  /// Epoch timestamp of when the price level was last updated
   final DateTime timestamp;
 
   Price(this.price, this.size, this.timestamp);