+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.1] - 2021-05-19
+
+* Fixed default DIO error handling
+* Added extra documentation to the API methods.
+
 ## [1.0.0] - 2021-05-10
 
 * Status